requirements.txt
python-telegram-bot==21.1.1
SQLAlchemy==2.0.23
boto3==1.28.78
botocore==1.31.78
psycopg2-binary==2.9.9
Pillow==10.3.0
httpx==0.27.0